+++ packages/doars-morph/src/directives/morph.js
@@ -24,25 +24,34 @@ export default {
         html = decode(html)
       }
 
-      // Ensure element only has one child.
-      if (element.children.length === 0) {
-        element.appendChild(document.createElement('div'))
-      } else if (element.children.length > 1) {
-        for (let i = element.children.length - 1; i >= 1; i--) {
-          element.children[i].remove()
+      if (modifiers.outer) {
+        // Morph the element as well.
+        morphTree(element, html)
+      } else {
+        // Ensure element only has one child.
+        if (element.children.length === 0) {
+          element.appendChild(document.createElement('div'))
+        } else if (element.children.length > 1) {
+          for (let i = element.children.length - 1; i >= 1; i--) {
+            element.children[i].remove()
+          }
         }
-      }
 
-      // Morph first child to given element tree.
-      const root = morphTree(element.children[0], html)
-      if (!element.children[0].isSameNode(root)) {
-        element.children[0].remove()
-        element.appendChild(root)
+        // Morph first child to given element tree.
+        const root = morphTree(element.children[0], html)
+        if (!element.children[0].isSameNode(root)) {
+          element.children[0].remove()
+          element.appendChild(root)
+        }
       }
     }
 
     // Execute value and retrieve result.
-    const result = processExpression(component, attribute, attribute.getValue())
+    const result = processExpression(
+      component,
+      attribute,
+      attribute.getValue(),
+    )
 
     // Store results.
     attribute.setData(result)