+++ packages/doars-navigate/dst/doars-navigate.iife.js
@@ -261,7 +261,6 @@
   };
 
   // src/factories/directives/navigate.js
-  var NAME = "navigate";
   var NAME_LOADER = "-loader";
   var NAME_TARGET = "-target";
   var HEADER_DATE = "Date";
@@ -309,10 +308,7 @@
     attribute[NAVIGATE].loaderTransitionIn = transitionIn(component, loaderElement);
   };
   var loaderRemove = (attribute, component, transitionOut) => {
-    if (attribute[NAVIGATE].loaderTransitionOut) {
-      return;
-    }
-    if (!attribute[NAVIGATE].loaderElement) {
+    if (attribute[NAVIGATE].loaderTransitionOut || !attribute[NAVIGATE].loaderElement) {
       return;
     }
     const loaderElement = attribute[NAVIGATE].loaderElement;
@@ -468,7 +464,7 @@
       });
     };
     return {
-      name: NAME,
+      name: "navigate",
       update: (component, attribute, {
         processExpression,
         transitionIn,
@@ -548,10 +544,22 @@
               }
             }
             if (modifiers.morph) {
-              html = fromString(html);
-              morphTree(target, html, {
-                childrenOnly: !modifiers.outer
-              });
+              if (modifiers.outer) {
+                morphTree(target, html);
+              } else {
+                if (target.children.length === 0) {
+                  target.appendChild(document.createElement("div"));
+                } else if (target.children.length > 1) {
+                  for (let i = target.children.length - 1; i >= 1; i--) {
+                    target.children[i].remove();
+                  }
+                }
+                const root = morphTree(target.children[0], html);
+                if (!target.children[0].isSameNode(root)) {
+                  target.children[0].remove();
+                  target.appendChild(root);
+                }
+              }
             } else if (modifiers.outer) {
               target.outerHTML = html;
             } else {
@@ -714,6 +722,7 @@
         }
         attribute[NAVIGATE] = {
           cache,
+          element,
           historyHandler,
           loadHandler,
           destroyPreloader
@@ -732,7 +741,7 @@
             cache = {};
           }
         }
-        attribute[NAVIGATE].target.removeEventListener(
+        attribute[NAVIGATE].element.removeEventListener(
           "click",
           attribute[NAVIGATE].loadHandler
         );