+++ packages/doars/dst/doars-call.iife.js
@@ -1456,8 +1456,14 @@
           return;
         }
         if (typeof html === "string") {
-          if (element.innerHTML !== html) {
-            element.innerHTML = html;
+          if (modifiers.outer) {
+            if (element.outerHTML !== html) {
+              element.outerHTML = html;
+            }
+          } else {
+            if (element.innerHTML !== html) {
+              element.innerHTML = html;
+            }
           }
           return;
         }