+++ packages/doars/dst/doars-interpret.js
@@ -1457,8 +1457,14 @@ var html_default = {
         return;
       }
       if (typeof html === "string") {
-        if (element.innerHTML !== html) {
-          element.innerHTML = html;
+        if (modifiers.outer) {
+          if (element.outerHTML !== html) {
+            element.outerHTML = html;
+          }
+        } else {
+          if (element.innerHTML !== html) {
+            element.innerHTML = html;
+          }
         }
         return;
       }