+++ packages/doars/src/directives/html.js
@@ -35,8 +35,14 @@ export default {
 
       // Set html via inner html for strings.
       if (typeof (html) === 'string') {
-        if (element.innerHTML !== html) {
-          element.innerHTML = html
+        if (modifiers.outer) {
+          if (element.outerHTML !== html) {
+            element.outerHTML = html
+          }
+        } else {
+          if (element.innerHTML !== html) {
+            element.innerHTML = html
+          }
         }
         return
       }