+++ packages/doars-navigate/CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG.md
 
+## 2.1.1
+
+- Fixed issue with morph modifier.
+
+## 2.1.0
+
+- Allow `loader` and `target` to return element selectors in addition to elements.
+
 ## 2.0.0
 
 - Initial release.