+++ packages/doars/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG.md
 
+## 2.1.0
+
+- Add outer modifier.
+
 ## 2.0.0
 
 - Added additional modifiers to the `on` directive.